+++ lib/trailblazer/developer/trace/present.rb
@@ -22,29 +22,47 @@ module Trailblazer::Developer
 
       # Entry point for rendering a {Trace::Stack}.
       # Used in `#wtf?`.
-      def call(stack, render_method: method(:render), node_options: {}, **options)
-        # The top activity doesn't have an ID, hence we need to compute a default label.
-        # TODO: maybe we should deep-merge here.
-        snapshot_before_for_top_activity = stack.to_a[0]
-
-        top_activity_options = {
-          # we can pass particular label "hints".
-          snapshot_before_for_top_activity => {
-            # label: %{#{snapshot_before_for_top_activity.task.superclass} (anonymous)},
-            label: snapshot_before_for_top_activity.task.inspect,
-          },
-        }
+      def call(stack, render_method: method(:render), node_options: nil, **options, &block)
+        raise "[Trailblazer] The `:node_options` option for `Trace::Present` is deprecated. Please use the block style as described here: #FIXME" if node_options
 
         # Build a generic array of {Trace::Node}s.
         trace_nodes = Trace.build_nodes(stack.to_a)
 
+        # The top activity doesn't have an ID, hence we need to compute a default label.
+        top_activity_trace_node = trace_nodes.to_a[0]
+
+        build_options = {
+          node_options: {
+            # we can pass particular label "hints".
+            top_activity_trace_node => {
+              # label: %{#{top_activity_trace_node.task.superclass} (anonymous)},
+              label: top_activity_trace_node.task.inspect,
+            },
+          }
+        }
+
+        build_options = merge_local_options(options, build_options)
+
         # specific rendering.
-        node_options = top_activity_options.merge(node_options)
+        options_from_block = block_given? ? block.call(trace_nodes: trace_nodes, stack: stack, **build_options) : {}
+        build_options = merge_local_options(options_from_block, build_options)
 
         # At this point we already decided that there is a Stack.
-        debugger_trace = Debugger::Trace.build(stack, trace_nodes, node_options: node_options, **options) # currently, we agree on using a Debugger::Node list as the presentation data structure.
+        debugger_trace = Debugger::Trace.build(stack, trace_nodes, **build_options) # currently, we agree on using a Debugger::Node list as the presentation data structure.
 
-        return render_method.(debugger_trace, **options)
+        return render_method.(debugger_trace, **build_options)
+      end
+
+      # @private
+      def merge_local_options(options, local_options)
+        merged_hash = options.collect do |key, value|
+          [
+            key,
+            value.is_a?(Hash) ? local_options.fetch(key, {}).merge(value) : value # options are winning over local_options[key]
+          ]
+        end.to_h
+
+        local_options.merge(merged_hash)
       end
     end # Present
   end