+++ test/trace_test.rb
@@ -43,12 +43,13 @@ class TraceTest < Minitest::Spec
   it "you can pass an explicit task label via {:label}" do
     stack, signal, (ctx, flow_options), _ = Dev::Trace.invoke(flat_activity, [{seq: []}, {}])
 
-    output = Dev::Trace::Present.(
-      stack,
-      node_options: {
-        stack.to_a[0] => {label: "#{flat_activity.class} (anonymous)"}
+    output = Dev::Trace::Present.(stack) do |trace_nodes:, **|
+      {
+        node_options: {
+          trace_nodes[0] => {label: "#{flat_activity.class} (anonymous)"}
+        }
       }
-    )
+    end
 
     assert_equal output, %{Class (anonymous)
 |-- Start.default
@@ -82,11 +83,14 @@ class TraceTest < Minitest::Spec
     assert_equal ctx_for.(stack.to_a[24], variable_versions), {:seq=>{:value=>"[:a, :b, :c, :d, :e]", :has_changed=>true}, :nil_value=>{:value=>"nil", :has_changed=>true}}
 
 # TODO: test label explicitely
-    output = Dev::Trace::Present.(stack,
-      node_options: {
-        stack.to_a[0] => {label: "#{activity.superclass} (anonymous)"},
+    output = Dev::Trace::Present.(stack) do |trace_nodes:, **|
+      {
+        node_options: {
+          trace_nodes[0] => {label: "#{activity.superclass} (anonymous)"},
+        }
       }
-    )
+    end
+
 
     puts output = output.gsub(/0x\w+/, "").gsub(/0x\w+/, "").gsub(/@.+_test/, "")
 
@@ -360,59 +364,5 @@ class TraceTest < Minitest::Spec
 
     assert_equal captured_input.data, { ctx: { seq: [:B, :C] }, something: :else }
     assert_equal captured_output.data, { ctx: { seq: [:B, :C] }, signal: signal }
-  end
-
-  it "{Present.call} accepts block to produce options that can be merged with original options" do
-    stack, signal, (ctx, flow_options), _ = Dev::Trace.invoke(flat_activity, [{seq: []}, {flow: true}])
-
-
-    output = Dev::Trace::Present.(stack,
-      node_options: {
-        stack.to_a[0] => {label: "<Anonymous activity>"}
-      }
-    )
-
-    assert_equal output, %{<Anonymous activity>
-|-- Start.default
-|-- B
-|-- C
-`-- End.success}
-  end
-
-  it "{Present.call} allows to inject :renderer and pass through additional arguments to the renderer (e.g. {:color})" do
-    stack, _ = Dev::Trace.invoke(nested_activity, [{ seq: [] }, {}])
-
-    renderer = ->(debugger_node:, tree:, color:, **) do
-      task = debugger_node.trace_node.snapshot_before.task
-
-      id_label = debugger_node.label
-
-      if task.is_a? Method
-        task = "#<Method: Trailblazer::Activity::Testing::Assertions::Implementing.#{task.name}>"
-      end
-      [
-        debugger_node.level,
-        %{#{debugger_node.level}/#{task}/#{debugger_node.trace_node.snapshot_after.data[:signal]}/#{id_label}/#{color}}
-      ]
-    end
-
-    output = Dev::Trace::Present.(
-      stack,
-      renderer: renderer,
-      color:    "pink", # additional options.
-    )
-
-    output = output.gsub(/0x\w+/, "").gsub(/0x\w+/, "").gsub(/@.+_test/, "")
-
-    _(output).must_equal %{0/#<Class:>/#<Trailblazer::Activity::End semantic=:success>/#<Class:>/pink
-|-- 1/#<Trailblazer::Activity::Start semantic=:default>/Trailblazer::Activity::Right/Start.default/pink
-|-- 1/#<Method: Trailblazer::Activity::Testing::Assertions::Implementing.b>/Trailblazer::Activity::Right/B/pink
-|-- 1/#<Class:>/#<Trailblazer::Activity::End semantic=:success>/D/pink
-|   |-- 2/#<Trailblazer::Activity::Start semantic=:default>/Trailblazer::Activity::Right/Start.default/pink
-|   |-- 2/#<Method: Trailblazer::Activity::Testing::Assertions::Implementing.B>/Trailblazer::Activity::Right/B/pink
-|   |-- 2/#<Method: Trailblazer::Activity::Testing::Assertions::Implementing.C>/Trailblazer::Activity::Right/C/pink
-|   `-- 2/#<Trailblazer::Activity::End semantic=:success>/#<Trailblazer::Activity::End semantic=:success>/End.success/pink
-|-- 1/#<Method: Trailblazer::Activity::Testing::Assertions::Implementing.e>/Trailblazer::Activity::Right/E/pink
-`-- 1/#<Trailblazer::Activity::End semantic=:success>/#<Trailblazer::Activity::End semantic=:success>/End.success/pink}
   end
 end